# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DeenMate API
  description: |
    Comprehensive Islamic application backend providing access to Quran, Hadith, Prayer Times, 
    Zakat calculations, and Islamic audio content. This API is designed to serve Islamic 
    applications, websites, and mobile apps with accurate, reliable, and well-structured Islamic data.
  
    ## Features
    - **Quran Module**: Complete Quran text, translations, and recitations
    - **Prayer Module**: Accurate prayer times, qibla direction, and calculation methods
    - **Hadith Module**: Authentic hadith collections and search functionality
    - **Zakat Module**: Zakat calculations and current market values
    - **Audio Module**: High-quality Islamic audio content and metadata
  
    ## Authentication
    Currently operates in read-only mode without authentication. Future versions will include 
    JWT-based authentication for user-specific features.
  
    ## Rate Limiting
    - Public Endpoints: 100 requests per minute per IP
    - Authenticated Endpoints: 1000 requests per minute per user
    - Premium Tier: 5000 requests per minute per user
  
    ## Support
    - Documentation: https://docs.deenmate.com
    - API Status: https://status.deenmate.com
    - Support Email: api-support@deenmate.com
    - Developer Discord: https://discord.gg/deenmate
  
  version: 1.0.0
  contact:
    name: DeenMate API Support
    email: api-support@deenmate.com
    url: https://docs.deenmate.com
  license:
...
servers:
...
paths:
...
  /zakat/nisab:
    get:
      tags:
        - Zakat
      summary: Get Nisab Values
      description: Retrieve current nisab values for zakat calculation
      parameters:
        - name: currency
          in: query
          description: Currency code
          schema:
            type: string
            default: USD
            enum: [USD, SAR, BDT, EUR, GBP]
      responses:
        '200':
          description: Current nisab values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NisabValuesResponse'
  /zakat/calculate:
    post:
      tags:
        - Zakat
      summary: Calculate Zakat
      description: Calculate zakat amount based on wealth details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateZakatRequest'
            example:
              gold: 1000
              silver: 0
              cash: 5000
              investments: 2000
              other: 0
              currency: "USD"
      responses:
        '200':
          description: Zakat calculation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZakatCalculationResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /zakat/gold-price:
    get:
      tags:
        - Zakat
      summary: Get Gold Price
      description: Retrieve current gold price
      parameters:
        - name: currency
          in: query
          description: Currency code
          schema:
            type: string
            default: USD
      responses:
        '200':
          description: Current gold price
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldPriceResponse'
...
  /zakat/history:
    get:
      tags:
        - Zakat
      summary: Get Zakat History
      description: Retrieve zakat calculation history
      parameters:
        - name: userId
          in: query
          description: User ID (for future authenticated access)
          schema:
            type: string
      responses:
        '200':
          description: Zakat calculation history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZakatHistoryResponse'
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          description: Response data
        meta:
          description: Metadata about the response
      required:
        - success
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            details:
              description: Additional error details
          required:
            - error
...
    CalculateZakatRequest:
      type: object
      properties:
        gold:
          type: number
          description: Gold value in specified currency
          default: 0
        silver:
          type: number
          description: Silver value in specified currency
          default: 0
        cash:
          type: number
          description: Cash value in specified currency
          default: 0
        investments:
          type: number
          description: Investment value in specified currency
          default: 0
        other:
          type: number
          description: Other assets value in specified currency
          default: 0
        currency:
          type: string
          description: Currency code
          default: USD
          enum: [USD, SAR, BDT, EUR, GBP]
      required:
        - currency
    ZakatCalculation:
      type: object
      properties:
        totalWealth:
          type: number
          description: Total wealth value
        zakatAmount:
          type: number
          description: Calculated zakat amount
        zakatPercentage:
          type: number
          description: Zakat percentage (2.5)
        nisabThreshold:
          type: number
          description: Nisab threshold value
        calculationDate:
          type: string
          format: date
        breakdown:
          type: object
          properties:
            gold:
              type: number
            silver:
              type: number
            cash:
              type: number
            investments:
              type: number
            other:
              type: number
    ZakatCalculationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              properties:
                calculation:
                  $ref: '#/components/schemas/ZakatCalculation'
            meta:
              type: object
              properties:
                calculationDate:
                  type: string
                  format: date
                nisabThreshold:
                  type: number
                zakatRate:
                  type: string

    # Audio Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints

# Global security (for future use)
# security:
#   - BearerAuth: []